// bootstrap4-course/src/app/services/router.enum.ts
export enum RouterEnum {
  BasicTypo = 'BasicTypo',
  TextDisplay = 'TextDisplay',
  FloatPosition = 'FloatPosition',
  ColorsBackGround = 'ColorsBackGround',
  Spacing = 'Spacing',
  Sizing = 'Sizing',
  BreakPoints = 'BreakPoints',
  Buttons = 'Buttons',
}

export enum HeaderMessageEnum {
  BasicTypo = 'Basic Typography',
  TextDisplay = 'Text Alignment & Display',
  FloatPosition = 'Floats & Fixed Positions',
  ColorsBackGround = 'Colors & Background',
  Spacing = 'Spacing',
  Sizing = 'Sizing & Borders',
  BreakPoints = 'Breakpoints',
  Buttons = 'Buttons & Button Groups',
}